created generic classroomManager policy and updated routes and related perms
