+++ cms/api/classroom-manager/config/routes.json
@@ -1,0 +1,13 @@
+{
+    "routes": [
+      {
+        "method": "GET",
+        "path": "/classroom-managers/me",
+        "handler": "classroom-manager.me",
+        "config": {
+          "policies": ["global::isClassroomManager"]
+        }
+      }
+    ]
+}
+
