+++ cms/api/mentor/config/routes.json
@@ -10,14 +10,6 @@
     },
     {
       "method": "GET",
-      "path": "/mentors/me",
-      "handler": "mentor.me",
-      "config": {
-        "policies": ["global::isMentor"]
-      }
-    },
-    {
-      "method": "GET",
       "path": "/mentors/count",
       "handler": "mentor.count",
       "config": {
@@ -37,7 +29,7 @@
       "path": "/mentors",
       "handler": "mentor.create",
       "config": {
-        "policies": []
+        "policies": ["global::isClassroomManager"]
       }
     },
     {
