+++ cms/api/teacher/config/routes.json
@@ -29,7 +29,7 @@
       "path": "/teachers",
       "handler": "teacher.create",
       "config": {
-        "policies": []
+        "policies": ["global::isClassroomManager"]
       }
     },
     {
